+ UserController, ReservasiController, belum diuji

+++ src/controller/SeasonController.ts
@@ -5,9 +5,14 @@ import PrismaScope from "../modules/PrismaService";
 import Validation from "../modules/Validation";
 import moment from "moment";
 import { Tarif } from "../modules/Models";
+import Authentication from "../modules/Authentication";
 
 export default class SeasonController {
     static async index(req: PegawaiRequest, res: Response) {
+        if(!Authentication.authorization(req, ['sm'])) {
+            return Authentication.defaultUnauthorizedResponse(res)
+        }
+
         const validation = Validation.query(req, {
             search: {
                 required: false
@@ -47,7 +52,42 @@ export default class SeasonController {
         })
     }
 
+    static async show(req: PegawaiRequest, res: Response) {
+        if(!Authentication.authorization(req, ['sm'])) {
+            return Authentication.defaultUnauthorizedResponse(res)
+        }
+
+        const { id } = req.params
+
+        return PrismaScope(async (prisma) => {
+            const season = await prisma.season.findUnique({
+                where: {
+                    id: +id
+                },
+                include: {
+                    tarif: true
+                }
+            })
+
+            if (season === null) {
+                return ApiResponse.error(res, {
+                    message: "Season tidak ditemukan",
+                    errors: null
+                }, 404)
+            }
+
+            return ApiResponse.success(res, {
+                message: "Berhasil mendapatkan data season",
+                data: season
+            })
+        })
+    }
+
     static async store(req: PegawaiRequest, res: Response) {
+        if(!Authentication.authorization(req, ['sm'])) {
+            return Authentication.defaultUnauthorizedResponse(res)
+        }
+
         const validation = Validation.body(req, {
             nama: {
                 required: true,
@@ -115,34 +155,11 @@ export default class SeasonController {
         })
     }
 
-    static async get(req: PegawaiRequest, res: Response) {
-        const { id } = req.params
-
-        return PrismaScope(async (prisma) => {
-            const season = await prisma.season.findUnique({
-                where: {
-                    id: +id
-                },
-                include: {
-                    tarif: true
-                }
-            })
-
-            if (season === null) {
-                return ApiResponse.error(res, {
-                    message: "Season tidak ditemukan",
-                    errors: null
-                }, 404)
-            }
-
-            return ApiResponse.success(res, {
-                message: "Berhasil mendapatkan data season",
-                data: season
-            })
-        })
-    }
-
     static async update(req: PegawaiRequest, res: Response) {
+        if(!Authentication.authorization(req, ['sm'])) {
+            return Authentication.defaultUnauthorizedResponse(res)
+        }
+
         const { id } = req.params
 
         const bodyValidation = Validation.body(req, {
@@ -210,10 +227,35 @@ export default class SeasonController {
     }
 
     static async destroy(req: PegawaiRequest, res: Response) {
+        if(!Authentication.authorization(req, ['sm'])) {
+            return Authentication.defaultUnauthorizedResponse(res)
+        }
+
         const { id } = req.params
 
         return PrismaScope(async (prisma) => {
-            const season = await prisma.season.delete({
+            const season = await prisma.season.findFirst({
+                where: {
+                    id: +id
+                }
+            })
+
+            if (season === null) {
+                return ApiResponse.error(res, {
+                    message: "Season tidak ditemukan",
+                    errors: null
+                }, 404)
+            }
+
+            // if tanggalStart < now + 2 bulan, return error
+            if (moment(season.tanggal_start).isBefore(SeasonController.getTanggalMaxInputSeason())) {
+                return ApiResponse.error(res, {
+                    message: "Season tidak dapat dihapus: tanggal mulai kurang dari 2 bulan dari sekarang",
+                    errors: null
+                }, 422)
+            }
+
+            await prisma.season.delete({
                 where: {
                     id: +id
                 }
@@ -234,7 +276,7 @@ export default class SeasonController {
 // Routing
 export const router = Router()
 router.get('/', SeasonController.index)
+router.get('/:id', SeasonController.show)
 router.post('/', SeasonController.store)
-router.get('/:id', SeasonController.get)
 router.put('/:id', SeasonController.update)
 router.delete('/:id', SeasonController.destroy)
